[ refactor ] A more concrete constraint for Sim
The previous constraint didn't dismiss the possibility of non standard
values (i.e. the space `[[ d ]] (Kripke V R) D i G` is larger than
just the values obtained by evaluating a term!). It was fine for our
purpose so far but in the POPLMark challenge we want the guarantee
that Sim indeed works on term.