[ done ] proof of simulation lemma after let-elab
I think this is a pretty good exposition of the annoying amount
of boilerplate that needs to be written before tackling the real
problem.

Additionally it's clear that ALL of the fusion proofs follow
exactly the pattern given in our generic library (and that makes
the proofs easy & the results very easy to use) except that here
it's hand-rolled.

I've managed to blaze-through this in about 4 intense hours only
because of the intimate knowledge of the way these things work.

It is also clear that this only gives us a proof about a very
specific system whereas let-elaboration as described in the paper
is general enough to apply to ALL possible target languages
(where source is defined as target + let).