[ upgrade ] Agda 2.6.0, stdlib 1.0
Re-using the refactorings I have done for my thesis to port the ICFP
paper. The idea being to be able to use this as a starting point for
the extended version and after to backport all of this to the library
itself too.

This is mostly working. Rare bits of the paper are still missing and
I need to turn off the sed scripts.