Compat 2.6.0 (#4)
* [ upgrade ] Agda 2.6.0, stdlib 1.0

Re-using the refactorings I have done for my thesis to port the ICFP
paper. The idea being to be able to use this as a starting point for
the extended version and after to backport all of this to the library
itself too.

This is mostly working. Rare bits of the paper are still missing and
I need to turn off the sed scripts.

* [ more ] adding back the combinators for indexed sets

* [ more ] almost there

A few things missing in Fusion still.

* [ done ] paper back to normal

* [ travis ] build using Agda 2.6.0 and stdlib 1.0

* [ admin ] move POPLMark examples

* [ refactor ] backport changes from src/ to doc/

* [ src ] extracted from doc/

* [ doc ] table of content

* [ fix ] missing src files

* [ fix ] hoping to get rid of the non-deterministic error

* Revert "[ fix ] hoping to get rid of the non-deterministic error"

This reverts commit 45e91c8a7a8839e054bebe42ae70b5d36d7df94a.
This is a known bug of 2.6.0 (cf. agda/agda#3694)